Rename evm.h to evmc.h
