+++ examples/capi.c
@@ -1,4 +1,4 @@
-#include <evm.h>
+#include <evmc.h>
 
 #include <inttypes.h>
 #include <string.h>
