+++ examples/examplevm.c
@@ -1,4 +1,4 @@
-#include <evm.h>
+#include <evmc.h>
 
 #include <limits.h>
 #include <stdio.h>
